requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
appdirs==1.4.4
arborize==6.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arviz==0.22.0
astropy==6.1.7
astropy-iers-data==0.2025.7.7.0.39.39
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blessed==1.21.0
bsb==6.0.0
bsb-core==6.0.0
bsb-hdf5==6.0.0
bsb-json==6.0.0
bsb-nest==6.0.0
bsb-neuron==6.0.0
bsb-yaml==6.0.0
cachetools==6.1.0
cerebellar-models==0.7.0
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cmake==3.31.6
comm==0.2.2
cons==0.4.7
contourpy==1.3.2
cookiecutter==2.6.0
corner==2.2.3
cycler==0.12.1
Cython==3.1.2
dashing-next==0.1.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
elephant==1.1.1
errr==1.2.0
etuples==0.3.10
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
find_libpython==0.4.0
fonttools==4.58.5
fsspec==2025.7.0
grpcio==1.74.0
h5netcdf==1.6.3
h5py==3.14.0
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
ipython-genutils==0.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
logical-unification==0.4.6
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
miniKanren==1.0.5
mistune==3.1.3
morphio==3.4.0
mpi4py==3.1.6
mpilock==1.1.0
mpipool==2.2.1
mpmath==1.3.0
multipledispatch==1.0.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo==0.13.3
nest-asyncio==1.6.0
NESTML==8.0.1
networkx==3.4.2
neuron-nightly==9.0a1.dev1422
nflows==0.14
nixio==1.5.4
nmodl-glia==6.0.0
notebook==6.4.12
nrn-patch==6.0.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
odetoolbox==2.5.8
opt_einsum==3.4.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyerfa==2.0.1.5
Pygments==2.19.2
pyknos==0.16.0
pymc==5.25.1
pynrrd==1.1.3
pyparsing==3.2.3
pyro-api==0.1.2
pyro-ppl==1.9.1
pytensor==2.31.7
pytest==8.4.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
quantities==0.15.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rtree==1.4.0
sbi==0.24.0
scikit-learn==1.7.0
scipy==1.15.3
Send2Trash==1.8.3
shortuuid==1.0.13
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
tblib==3.1.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.5.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xarray==2025.6.1
xarray-einstats==0.8.0
zipp==3.23.0
zuko==1.4.1
